Update inviewport.jquery.js
Updated script to not only add a class when in view, it will check if it is in view wither your scrolling vertically or horizontally and it will also trigger a function call outside this script for use with things like animations. Checking if an animation is in view and starting the animation then and turning it off when out of view for example. You can call a function once or every iteration. Example code provided.